requirements.txt
Django==3.1.4
paramiko==2.4.3
python-nmap==0.6.1
pymysql==0.10.1